:truck: Move "submit to supervisor" button into approval form card namespace.
Why? to associate it with the component its being used in.
